% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/above_percent.R
\name{above_percent}
\alias{above_percent}
\title{Calculate percentage of values above target thresholds}
\usage{
above_percent(data, targets_above = c(60, 85, 100))
}
\arguments{
\item{data}{A DataFrame object with column names "id", "time", "hr".
Missing HR values (NA) are automatically excluded from calculations.}

\item{targets_above}{\strong{Default: (60, 85, 100).} Numeric vector of heart rate thresholds. Heart rate values from
data argument will be compared to each value in the targets_above vector.}
}
\value{
If a DataFrame object is passed, then a tibble object with
a column for subject id and then a column for each target value is returned. If a vector of heart rate
values is passed, then a tibble object without the subject id is returned. Wrap
\code{as.numeric()} around the latter to output a numeric vector.
}
\description{
The function \code{above_percent} produces a tibble object with values equal to
the percentage of heart rate data above target values. The output columns
correspond to the subject id followed by the target values, and the
output rows correspond to the subjects. The values will be between 0
(no measurements) and 100 (all measurements).
}
\details{
A tibble object with 1 row for each subject, a column for subject id and
column for each target value is returned. NA's will be omitted from the glucose
values in calculation of percent.
}
\examples{

data(example_heart_1)

above_percent(example_heart_1)
above_percent(example_heart_1, targets_above = c(100, 150, 180))

}
