% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_HRR.R
\name{calculate_HRR}
\alias{calculate_HRR}
\title{Calculate heart rate reserve}
\usage{
calculate_HRR(data, method, rhr_method, quantile_val, tz)
}
\arguments{
\item{data}{A DataFrame object with column names "id", "time", "hr".}

\item{method}{\strong{Default: "max-RHR".} A choice for the user to choose which formula they want to use: HRR = max - Resting heart rate ("max-RHR") or HRR = max - min ("max-min")}

\item{rhr_method}{\strong{Default: "mean".} Refer to the documentation for \code{\link{calculate_RHR}} for more details.}

\item{quantile_val}{A choice for the user to choose which quantile value they want to use(most common 1, if there's strange max value, recommend to use 0.99)}

\item{tz}{A character string specifying the time zone to be used. System-specific (see \code{\link{as.POSIXct}}), but " " is the current time zone, and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC, on some platforms with a warning}
}
\value{
If a dataframe object is passed, then a tibble object with a column for subject id and a column for each of summary values is returned.
'NA' heartrate values are omitted from the calculation of the summary values.
}
\description{
Computes the heart rate reserve (HRR), defined as the difference between the maximum heart rate and resting heart rate.
For details on computing resting heart rate, see \code{\link{calculate_RHR}}
}
\examples{
data(example_heart_1)
calculate_HRR(example_heart_1, 'max-min')
calculate_HRR(example_heart_1, quantile_val = 0.99)
}
\references{
Heart rate reserve as a predictor of cardiovascular and all-cause mortality in men,
\doi{10.1249/01.MSS.0000038959.67619.30}.
}
