% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicalc_profile.R
\name{epicalc_profile}
\alias{epicalc_profile}
\title{Display Episode Calculation statistics for selected subject}
\usage{
epicalc_profile(
  data,
  dur_length = 15,
  end_length = 15,
  subject = NULL,
  dt0 = 1,
  inter_gap = 15,
  tz = ""
)
}
\arguments{
\item{data}{DataFrame object with column names "id", "time", "hr"}

\item{dur_length}{\strong{Default: 15.} Numeric value specifying the minimum duration in minutes to be
considered an episode. Note dur_length should be a multiple of the data recording
interval otherwise the function will round up to the nearest multiple.}

\item{end_length}{\strong{Default: 15.} Numeric value specifying the minimum duration in minutes of
improved heart rate for an episode to end.}

\item{subject}{String corresponding to subject id}

\item{dt0}{Grid Length}

\item{inter_gap}{The maximum allowable gap (in minutes) for interpolation.}

\item{tz}{A character string specifying the time zone to be used. System-specific (see \code{\link{as.POSIXct}}), but " " is the current time zone, and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC, on some platforms with a warning}
}
\value{
A plot displaying (1) the statistics for the episodes and (2) the episodes colored by level.
}
\description{
Display Episode Calculation statistics for selected subject
}
\details{
For details on episode calculation, see \code{\link{episode_calculation}}
}
\examples{
epicalc_profile(example_heart_1)

}
