% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_calculation.R
\name{episode_calculation}
\alias{episode_calculation}
\title{Calculates low/high heartrate episodes with summary statistics}
\usage{
episode_calculation(
  data,
  lv1_low = 60,
  lv2_low = 55,
  lv1_high = 85,
  lv2_high = 100,
  dur_length = 15,
  end_length = 15,
  return_data = FALSE,
  dt0 = NULL,
  inter_gap = 45,
  tz = ""
)
}
\arguments{
\item{data}{DataFrame object with column names "id", "time", "hr"}

\item{lv1_low}{Numeric value specifying a low heart rate threshold for level 1}

\item{lv2_low}{Numeric value specifying a low heart rate threshold for level 2}

\item{lv1_high}{Numeric value specifying a high heart rate threshold for level 1}

\item{lv2_high}{Numeric value specifying a high heart rate threshold for level 2}

\item{dur_length}{Numeric value specifying the minimum duration in minutes to be
considered an episode. Note dur_length should be a multiple of the data recording
interval otherwise the function will round up to the nearest multiple. Default
is 15 minutes to match consensus.}

\item{end_length}{Numeric value specifying the minimum duration in minutes of
improved heart rate for an episode to end. Default is equal to dur_length to match consensus.}

\item{return_data}{Boolean indicating whether to also return data with episode labels.
Defaults to FALSE which means only episode summary statistics will be returned}

\item{dt0}{Grid Length}

\item{inter_gap}{The maximum allowable gap (in minutes) for interpolation.}

\item{tz}{A character string specifying the time zone to be used. System-specific (see \code{\link{as.POSIXct}}), but " " is the current time zone, and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC, on some platforms with a warning}
}
\value{
If return_data is FALSE, a single dataframe with columns:
\item{id}{Subject id}
\item{type}{Type of episode - either low or high}
\item{level}{Level of episode - one of lv1, lv2, extended, lv1_excl}
\item{avg_ep_per_day}{Average number of episodes per day calculated as
(total # episodes)/(recording time in days (24hrs))}
\item{avg_ep_duration}{Average duration of episodes in minutes}
\item{avg_ep_hr}{Average heart rate in the episode in minutes}
\item{total_episodes}{Total number of episodes in the subject's heart rate trace}

If return_data is TRUE, returns a list where the first entry is the episode summary dataframe
(see above) and the second entry is the input data with episode labels added. Note
the data returned here has been interpolated using the HR2DayByDay() function.
Mostly for use with epicalc_profile function. Format of the second list entry is:
\item{id}{Subject id}
\item{time}{Interpolated timestamps}
\item{hr}{heart rate in minutes}
}
\description{
The function determines episodes or events, calculates summary statistics,
and optionally returns data with episode label columns added
}
\details{
Note we have classified lv2 as a subset of lv1 since we find the consensus to be
slightly ambiguous. For lv1 exclusive of lv2, please see lv1_excl which summarises
episodes that were exclusively lv1 and did not cross the lv2 threshold. Also note,
low extended refers to episodes that are >120 consecutive minutes below lv1 low
and ends with at least 15 minutes of normal heart rate.
}
\examples{
episode_calculation(example_heart_1, lv1_low = 100, lv1_high = 120)

}
\references{
“What Your Heart Rate Is Telling You.” Harvard Health, 13 June 2023,
www.health.harvard.edu/heart-health/what-your-heart-rate-is-telling-you.
}
