% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_heart_1}
\alias{example_heart_1}
\title{Example heart rate data set for one subject}
\format{
A data frame with 154104 rows and 3 variables:
\describe{
\item{id}{id, as provided by Kaggle}
\item{time}{Time of observation, in format \%Y-\%m-\%d \%H:\%M:\%S}
\item{hr}{Heart rate, measured in BPM}
}
}
\source{
\url{https://www.kaggle.com/datasets/arashnic/fitbit/versions/1}
}
\usage{
example_heart_1
}
\description{
FitBit heart rate measurement from one subject
}
\keyword{datasets}
