% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_heart_5}
\alias{example_heart_5}
\title{Example heart rate data set for five subjects}
\format{
A data frame with 530859 rows and 3 variables:
\describe{
\item{id}{id, as provided by Kaggle}
\item{time}{time of observation, in format \%Y-\%m-\%d \%H:\%M:\%S}
\item{hr}{heart rate, measured in BPM}
}
}
\source{
\url{https://www.kaggle.com/datasets/arashnic/fitbit/versions/1}
}
\usage{
example_heart_5
}
\description{
FitBit heart rate measurement from five subjects
}
\keyword{datasets}
