% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_roc.R
\name{hist_roc}
\alias{hist_roc}
\title{Plot histogram of Rate of Change values (ROC)}
\usage{
hist_roc(data, timelag = 1, inter_gap = 15, tz = "")
}
\arguments{
\item{data}{DataFrame object with column names "id", "time", "hr"}

\item{timelag}{Integer indicating the time period (# minutes) over which rate
of change is calculated. Default is 1, e.g. rate of change is the change in
heart rate over the past 1 minute divided by 1}

\item{inter_gap}{\strong{Default: 15.} The maximum allowable gap (in minutes) for interpolation. The values will not be interpolated between the glucose measurements that are more than inter_gap minutes apart.}

\item{tz}{A character string specifying the time zone to be used. System-specific (see \code{\link{as.POSIXct}}), but " " is the current time zone, and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC, on some platforms with a warning}
}
\description{
The function hist_roc produces a histogram plot of ROC values
}
\details{
For additional information on calculating ROC, see \code{\link{roc}}
}
\examples{
data(example_heart_1)
hist_roc(example_heart_1)
}
