% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahe.R
\name{mahe}
\alias{mahe}
\title{Calculate Mean Amplitude of Heartrate Excursions}
\usage{
mahe(
  data,
  short_ma = 5,
  long_ma = 32,
  return_type = c("num", "df"),
  direction = c("avg", "service", "max", "plus", "minus"),
  tz = "",
  inter_gap = 45,
  max_gap = 180,
  plot = FALSE,
  title = NA,
  xlab = NA,
  ylab = NA,
  show_ma = FALSE,
  show_excursions = TRUE
)
}
\arguments{
\item{data}{DataFrame object with column names "id", "time", "hr"}
}
\value{
A tibble object with two columns: the subject id and corresponding MAHE value.
}
\description{
The function calculates MAHE values (a metric inspired by MAGE from glucose monitoring) and optionally returns a heart trace plot
}
\references{
Service et al. (1970) Mean amplitude of glycemic excursions, a measure of diabetic instability
\emph{Diabetes}  \strong{19} .644-655,
\doi{10.2337/diab.19.9.644}.

Fernandes, Nathaniel J., et al. "Open-source algorithm to calculate mean amplitude of glycemic excursions using short and long moving averages."
Journal of diabetes science and technology 16.2 (2022): 576-577. \doi{10.1177/19322968211061165}
}
