% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missingness_by_period.R
\name{missingness_by_period}
\alias{missingness_by_period}
\title{Calculate the \% missingness of the heart rate value during different time period}
\usage{
missingness_by_period(data, tz)
}
\arguments{
\item{data}{A DataFrame object with column names "id", "time", "hr".}

\item{tz}{A character string specifying the time zone to be used. System-specific (see \code{\link{as.POSIXct}}), but " " is the current time zone, and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC, on some platforms with a warning
Missing HR values (NA) are automatically excluded from calculations.}
}
\value{
If a dataframe object is passed, then a tibble object with a column for subject id and a column for each of percentage missingness is returned.
'NA' heartrate values are omitted from the calculation of the summary values.
}
\description{
The function missingness_by_period computes the percentage missingness of the heart rate data during six different time periods each day.
Choose 3am-7am as one of the time period due to the definition of resting heart rate in the reference. The other time periods are set following this 3am-7am period.
}
\examples{
data(example_heart_1)
missingness_by_period(example_heart_1)
}
