% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_nonmissing.R
\name{percent_nonmissing}
\alias{percent_nonmissing}
\title{Calculate the \% missingness of the heart rate value}
\usage{
percent_nonmissing(data)
}
\arguments{
\item{data}{A DataFrame object with column names "id", "time", "hr".
Missing HR values (NA) are automatically excluded from calculations.}
}
\value{
If a dataframe object is passed, then a tibble object with a column for subject id and a column for each of percentage missingness is returned.
'NA' heartrate values are omitted from the calculation of the summary values.
}
\description{
The function percent_nonmissing computes the percentage missingness of the heart rate data.
}
\examples{
data(example_heart_1)
percent_nonmissing(example_heart_1)
}
