% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ahp.R
\name{plot_ahp}
\alias{plot_ahp}
\title{Plot Ambulatory Heart Rate Profile (AHP) modal day}
\usage{
plot_ahp(data, smooth = TRUE, span = 0.3,
inter_gap = 15, tz = "", title = FALSE)
}
\arguments{
\item{data}{DataFrame object with column names "id", "time", "hr"}

\item{smooth}{Boolean indicating whether quantiles should be smoothed before plotting, default is TRUE}

\item{span}{Optional parameter indicating span for loess smoothing. Default is 0.3, larger values result in more smoothing,
recommended to choose between 0.1 to 0.7.}

\item{inter_gap}{The maximum allowable gap (in minutes) for interpolation. The values will not be interpolated between the Heart Rate measurements that are more than inter_gap minutes apart.}

\item{tz}{A character string specifying the time zone to be used. System-specific (see \code{\link{as.POSIXct}}), but " " is the current time zone, and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC, on some platforms with a warning}

\item{title}{Indicator whether the title of the plot should display the subject ID. The default is FALSE (no title).}
}
\value{
Plot of a 24 hr modal day collapsing all data to a single day.
}
\description{
The function plot_ahp produces an AhP plot that collapses all data into a single 24 hr "modal day".
}
\examples{
data(example_heart_1)
plot_ahp(example_heart_1)

}
