% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hr.R
\name{plot_hr}
\alias{plot_hr}
\title{Plotting time series plot of heart rate measurements}
\usage{
plot_hr(
  data,
  LHR = 60,
  UHR = 100,
  from = "",
  to = "",
  agg = c("none", "minute", "hour"),
  inter_gap = 60,
  tz = ""
)
}
\arguments{
\item{data}{Dataframe with column names ("id, "time", "hr")}

\item{LHR}{Lower Heart Rate set at \strong{default = 60}}

\item{UHR}{Upper Heart Rate set at \strong{default = 100}}

\item{from}{Starting time of plot inclusive (In format \%Y-\%m-\%d)}

\item{to}{Ending time of plot exclusive (In format \%Y-\%m-\%d)}

\item{agg}{Aggregate data by specified time interval}

\item{inter_gap}{Gap allowed between consecutive observations (in seconds)}

\item{tz}{A character string specifying the time zone to be used. System-specific (see \code{\link{as.POSIXct}}), but " " is the current time zone, and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC, on some platforms with a warning}
}
\value{
Heart rate time series plot for a single subject
}
\description{
The function 'plot_hr' supports time series plotting for both single and multiple subject data
}
\details{
See reasoning for default LHR and UHR values at \url{https://www.heart.org/en/health-topics/arrhythmia/about-arrhythmia/tachycardia--fast-heart-rate}
}
\examples{
data(example_heart_1)
plot_hr(example_heart_1)

}
