% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ranges_PA.R
\name{plot_ranges_PA}
\alias{plot_ranges_PA}
\title{Plot Time in Ranges as a bar plot}
\usage{
plot_ranges_PA(data, tz)
}
\arguments{
\item{data}{A DataFrame object with column names "id", "time", "hr".}

\item{tz}{A character string specifying the time zone to be used. System-specific (see \code{\link{as.POSIXct}}), but " " is the current time zone, and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC, on some platforms with a warning
Missing HR values (NA) are automatically excluded from calculations.}
}
\value{
Single subject bar chart showing percent in different glucose ranges.
}
\description{
The function plot_ranges produces a barplot showing the percent of time in different stages of heart rate
}
\details{
Only a single subject's data may be used. sedentary/sleep (<20\% HRR), light (20-39\% HRR), moderate (40-59\% HRR), and vigorous (>=60\% HRR). For details on HRR calculations see \code{\link{calculate_HRR}}
This plot is meant to be used as part of the Ambulatory Heart rate Profile (AHP)
}
\examples{

data(example_heart_1)
plot_ranges_PA(example_heart_1)

}
\references{
Using Heart Rate and Accelerometry to Define Quantity and Intensity of Physical Activity in Older Adults
\doi{10.1093/gerona/gly029}
}
