% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_roc.R
\name{scatter_roc}
\alias{scatter_roc}
\title{Plot the relationship between Rate of Change (ROC) and Heart Rate}
\usage{
scatter_roc(data, timelag = 1, inter_gap = 15, tz = "")
}
\arguments{
\item{data}{DataFrame object with column names "id", "time", "hr"}

\item{timelag}{Integer indicating the time period (# minutes) over which rate
of change is calculated. Default is 1, e.g. rate of change is the change in
heart rate over the past 1 minute divided by 1}

\item{inter_gap}{\strong{Default: 15.} The maximum allowable gap (in minutes) for interpolation. The values will not be interpolated between the glucose measurements that are more than inter_gap minutes apart.}

\item{tz}{A character string specifying the time zone to be used. System-specific (see \code{\link{as.POSIXct}}), but " " is the current time zone, and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC, on some platforms with a warning}
}
\value{
A ggplot object with ROC on the X-axis and the latter Heart Rate value on the Y-axis
}
\description{
The function \code{scatter_roc} produces a scatterplot of ROC and Heart Rate
}
\details{
A ggplot is returned that shows the relationship of a subject's Rate of Change and the subject's
heart rate. This allows a personalized insight into every subject's ROC, understanding the heart rate in which they ended at
and whether the ROC value comes from a higher activity heart rate or from another context.

The horizontal lines separate heart rate into 4 distinct sections. These are similar to that generated from
\code{plot_ranges_PA}. The thresholds here are as follows:
\enumerate{
\item Vigorous: >= 60 \% HRR + RHR
\item Moderate: 40 \% HRR + RHR -  60 \% HRR + RHR
\item Light: 20 \% HRR + RHR -  40 \% HRR + RHR
\item Sedentary/Sleep: <= 20 \% HRR + RHR
}
}
\examples{
data(example_heart_1)
scatter_roc(example_heart_1)

}
