% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_PA.R
\name{summarize_PA}
\alias{summarize_PA}
\title{Calculate the time an individual spend in different stages(following the definition in the reference)}
\usage{
summarize_PA(data, tz)
}
\arguments{
\item{data}{A DataFrame object with column names "id", "time", "hr".}

\item{tz}{A character string specifying the time zone to be used. System-specific (see \code{\link{as.POSIXct}}), but " " is the current time zone, and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC, on some platforms with a warning
Missing HR values (NA) are automatically excluded from calculations.}
}
\value{
If a dataframe object is passed, then a tibble object with a column for subject id and a column for each of summary values is returned.
'NA' heartrate values are omitted from the calculation of the summary values.
}
\description{
The function summary_PA computes the time(in percentage) for different individuals spend in different stages
Using the thresholds: sedentary/sleep (<20\% HRR), light (20–39\% HRR), moderate (40–59\% HRR), and vigorous (≥60\% HRR)
}
\examples{
data(example_heart_1)
summarize_PA(example_heart_1)
}
\references{
Using Heart Rate and Accelerometry to Define Quantity and Intensity of Physical Activity in Older Adults
\doi{10.1093/gerona/gly029}
}
