% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_hr.R
\name{summary_hr}
\alias{summary_hr}
\title{Calculate summary heart rate}
\usage{
summary_hr(data)
}
\value{
If a dataframe object is passed, then a tibble object with a column for subject id and a column for each of summary values is returned.
'NA' heartrate values are omitted from the calcution of the summary values.
}
\description{
Calculate summary heart rate
}
\examples{
data(example_heart_1)
summary_hr(example_heart_1)
}
