% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_days.R
\name{valid_days}
\alias{valid_days}
\title{Boolean value for a valid day of HR data}
\usage{
valid_days(data, hour = 10, min_in_hr = 60)
}
\arguments{
\item{data}{Dataframe with column names ("id, "time", "hr")}

\item{hour}{Hour value that a day must have in order for that corresponding day to be considered valid. \strong{Default: 10.}}

\item{min_in_hr}{The number of minutes within the hour that an observation must have for an hour to contribute to a valid day. \strong{Default: 60.}}
}
\value{
A tibble item with the columns id, day, full_hours, valid_day
}
\description{
The function 'valid_days' computes whether a day within an individual's HR data is valid, defined as
whether a day has more than a pre-specified hours worth of data.
}
\details{
For an individual's HR data, steps must be taken to ensure the quality of the data that is analyzed. 'valid_days' is
designed to return the days in which an individual has worn their device for a sufficient amount for unbiased estimates
(To the discretion of the data analyst). Default values selected by most common approaches (Reference Provided)
}
\examples{
data(example_heart_1)
valid_days(example_heart_1)

}
\references{
Reporting adherence, validity and physical activity measures of wearable activity trackers in medical research: A systematic review
\doi{10.1016/j.ijmedinf.2022.104696}
}
