% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/active_percent.R
\name{active_percent}
\alias{active_percent}
\title{Calculate the \% missingness of the heart rate value}
\usage{
active_percent(data)
}
\arguments{
\item{data}{A DataFrame  mobject with column names "Id", "Time", "Value".
Missing HR values (NA) are automatically excluded from calculations.}
}
\value{
If a dataframe object is passed, then a tibble object with a column for subject id and a column for each of percentage missingness is returned.
'NA' heartrate values are omitted from the calcution of the summary values.
}
\description{
The function active_percent computes the percentage missingness of the heart rate data.
}
\examples{
data(example_heart_1)
active_percent(example_heart_1)
}
