% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ahp.R
\name{ahp}
\alias{ahp}
\title{Display Ambulatory Heart rate Profile (AHP) statistics for selected subject}
\usage{
ahp(data, maxd = 14, inter_gap = 15, dt0 = 1, tz = "", daily = TRUE)
}
\arguments{
\item{data}{DataFrame object with column names "id", "time", "hr"}

\item{maxd}{max days}

\item{inter_gap}{The maximum allowable gap (in minutes) for interpolation.}

\item{dt0}{Grid Length}

\item{tz}{A character string specifying the time zone to be used. System-specific (see \code{\link{as.POSIXct}}), but " " is the current time zone, and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC, on some platforms with a warning}

\item{daily}{\strong{Default: TRUE.} Logical indicator whether AHP should include separate daily plots.}
}
\value{
A plot displaying heart rate measurements range, selected heart rate statistics (Resting Heart Rate, Heart Rate Reserve), percentage spent in target ranges and quantiles of 24 hour profile.
}
\description{
Display Ambulatory Heart rate Profile (AHP) statistics for selected subject
}
\details{
For details on the different metrics and plots see the following:
Time in range bar plot (see \code{\link{plot_ranges_PA}})
Summary quantile plot (see \code{\link{plot_ahp}})
Daily Plot (see \code{\link{plot_daily}})
Resting Heart Rate (see \code{\link{calculate_RHR}})
Heart Rate Reserve (see \code{\link{calculate_HRR}})
Range of Measurement Data (see \code{\link{percent_nonmissing}})
}
\examples{
data(example_heart_1)
ahp(example_heart_1, daily = FALSE)
}
