% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_HRR.R
\name{calculate_HRR}
\alias{calculate_HRR}
\title{Calculate heart rate reserve}
\usage{
calculate_HRR(data)
}
\arguments{
\item{data}{A DataFrame object with column names "id", "time", "hr".}
}
\value{
If a dataframe object is passed, then a tibble object with a column for subject id and a column for each of summary values is returned.
'NA' heartrate values are omitted from the calculation of the summary values.
}
\description{
The function calculate_RHR computes the heart rate reserve.
It calculates HRR using the formula:maximum heart rate - resting heart rate following the definition gained from the reference.
}
\examples{
data(example_heart_1)
calculate_HRR(example_heart_1)
}
\references{
Heart rate reserve as a predictor of cardiovascular and all-cause mortality in men,
\doi{10.1249/01.MSS.0000038959.67619.30}.
}
