% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_RHR.R
\name{calculate_RHR}
\alias{calculate_RHR}
\title{Calculate resting heart rate}
\usage{
calculate_RHR(data)
}
\arguments{
\item{data}{A DataFrame object with column names "id", "time", "hr".}
}
\value{
If a dataframe object is passed, then a tibble object with a column for subject id and a column for each of summary values is returned.
'NA' heartrate values are omitted from the calculation of the summary values.
}
\description{
The function calculate_RHR computes the resting heart rate.
It calculates the mean hr value between 3am to 7am and consider it as the resting heart rate according to the reference.
}
\examples{
data(example_heart_1)
summary_hr(example_heart_1)
}
\references{
Measure by measure: Resting heart rate across the 24-hour cycle,
\doi{10.1371/journal.pdig.0000236}.
}
