% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_RHR.R
\name{calculate_RHR}
\alias{calculate_RHR}
\title{Calculate resting heart rate}
\usage{
calculate_RHR(data, tz)
}
\arguments{
\item{data}{A DataFrame object with column names "id", "time", "hr".}

\item{tz}{A character string specifying the time zone to be used. System-specific (see \code{\link{as.POSIXct}}), but " " is the current time zone, and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC, on some platforms with a warning}
}
\value{
If a dataframe object is passed, then a tibble object with a column for subject id and a column for each of summary values is returned.
'NA' heartrate values are omitted from the calculation of the summary values.
}
\description{
Computes the resting heart rate (RHR) according to the mean heart rate
value between 3am to 7am. See reference for details on computing RHR
}
\examples{
data(example_heart_1)
calculate_RHR(example_heart_1)
}
\references{
Measure by measure: Resting heart rate across the 24-hour cycle,
\doi{10.1371/journal.pdig.0000236}.
}
