% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epicalc_profile.R
\name{epicalc_profile}
\alias{epicalc_profile}
\title{Display Episode Calculation statistics for selected subject}
\usage{
epicalc_profile(
  data,
  lv1_low = 60,
  lv2_low = 55,
  lv1_high = 85,
  lv2_high = 100,
  dur_length = 15,
  end_length = 15,
  subject = NULL,
  dt0 = NULL,
  inter_gap = 45,
  tz = ""
)
}
\arguments{
\item{data}{DataFrame object with column names "id", "time", "hr"}

\item{lv1_low}{Numeric value specifying a low heart rate threshold for level 1}

\item{lv2_low}{Numeric value specifying a low heart rate threshold for level 2}

\item{lv1_high}{Numeric value specifying a high heart rate threshold for level 1}

\item{lv2_high}{Numeric value specifying a high heart rate threshold for level 2}

\item{dur_length}{Numeric value specifying the minimum duration in minutes to be
considered an episode. Note dur_length should be a multiple of the data recording
interval otherwise the function will round up to the nearest multiple. Default
is 15 minutes to match consensus.}

\item{end_length}{Numeric value specifying the minimum duration in minutes of
improved heart rate for an episode to end. Default is equal to dur_length to match consensus.}

\item{subject}{String corresponding to subject id}

\item{dt0}{Grid Length}

\item{inter_gap}{The maximum allowable gap (in minutes) for interpolation.}

\item{tz}{A character string specifying the time zone to be used. System-specific (see \code{\link{as.POSIXct}}), but " " is the current time zone, and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC, on some platforms with a warning}
}
\value{
A plot displaying (1) the statistics for the episodes and (2) the episodes colored by level.
}
\description{
Display Episode Calculation statistics for selected subject
}
\examples{
epicalc_profile(example_heart_1)

}
\seealso{
episode_calculation()
}
