% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/episode_calculation.R
\name{episode_calculation}
\alias{episode_calculation}
\title{Calculates low/high heartrate episodes with summary statistics}
\usage{
episode_calculation(
  data,
  dur_length = 15,
  end_length = 15,
  return_data = FALSE,
  dt0 = 1,
  inter_gap = 15,
  tz = ""
)
}
\arguments{
\item{data}{DataFrame object with column names "id", "time", "hr"}

\item{dur_length}{Numeric value specifying the minimum duration in minutes to be
considered an episode. Note dur_length should be a multiple of the data recording
interval otherwise the function will round up to the nearest multiple. Default
is 15 minutes to match consensus.}

\item{end_length}{Numeric value specifying the minimum duration in minutes of
improved heart rate for an episode to end. Default is equal to dur_length to match consensus.}

\item{return_data}{Boolean indicating whether to also return data with episode labels.
Defaults to FALSE which means only episode summary statistics will be returned}

\item{dt0}{Grid Length}

\item{inter_gap}{The maximum allowable gap (in minutes) for interpolation.}

\item{tz}{A character string specifying the time zone to be used. System-specific (see \code{\link{as.POSIXct}}), but " " is the current time zone, and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC, on some platforms with a warning}
}
\value{
If return_data is FALSE, a single dataframe with columns:
\item{id}{Subject id}
\item{type}{Type of episode - either low or high}
\item{level}{Level of episode - one of lv1, lv2, extended, lv1_excl}
\item{avg_ep_per_day}{Average number of episodes per day calculated as
(total # episodes)/(recording time in days (24hrs))}
\item{avg_ep_duration}{Average duration of episodes in minutes}
\item{avg_ep_hr}{Average heart rate in the episode in minutes}
\item{total_episodes}{Total number of episodes in the subject's heart rate trace}

If return_data is TRUE, returns a list where the first entry is the episode summary dataframe
(see above) and the second entry is the input data with episode labels added. Note
the data returned here has been interpolated using the HR2DayByDay() function.
Mostly for use with epicalc_profile function. Format of the second list entry is:
\item{id}{Subject id}
\item{time}{Interpolated timestamps}
\item{hr}{heart rate in minutes}
}
\description{
The function determines episodes or events, calculates summary statistics,
and optionally returns data with episode label columns added
}
\details{
Using the thresholds: sedentary/sleep (<20\% HRR), light (20–39\% HRR), moderate (40–59\% HRR), and vigorous (≥60\% HRR)
}
\examples{
episode_calculation(example_heart_1)

}
\references{
Using Heart Rate and Accelerometry to Define Quantity and Intensity of Physical Activity in Older Adults
\doi{10.1093/gerona/gly029}
}
