% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_daily.R
\name{plot_daily}
\alias{plot_daily}
\title{Plot daily heart rate profiles}
\usage{
plot_daily(data, maxd = 14, inter_gap = 15, tz = "")
}
\arguments{
\item{data}{DataFrame object with column names "id", "time", "hr"}

\item{maxd}{\strong{Default: 14.} Number of days to plot. If less than \code{maxd} days of data are available, all days are plotted.}

\item{inter_gap}{The maximum allowable gap (in minutes) for interpolation.}

\item{tz}{A character string specifying the time zone to be used. System-specific (see \code{\link{as.POSIXct}}), but " " is the current time zone, and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC, on some platforms with a warning}
}
\value{
Daily heart rate time series plots for a single subject
}
\description{
The function \code{plot_daily} plots daily heart rate time series profiles for a single subject.
}
\details{
Only a single subject's data may be plotted. The black line shows the heart rate values.
The shaded orange area shows the moderate range, The shaded green area shows the light range, Areas of the curve
above the hr_60 are shaded red(which is the vigorous range), while areas below the hr_20 are shaded blue(which is the Sedentary/Sleep range).
}
\examples{

data(example_heart_1)
plot_daily(example_heart_1)

}
