% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hr.R
\name{plot_hr}
\alias{plot_hr}
\title{Plotting time series plot of heart rate measurements}
\usage{
plot_hr(
  data,
  LHR = 60,
  UHR = 100,
  from = "",
  to = "",
  agg = c("none", "minute", "hour"),
  inter_gap = 60
)
}
\arguments{
\item{data}{Dataframe with column names ("id, "time", "hr")}

\item{LHR}{Lower heart rate set at \strong{default = 60 BPM}}

\item{UHR}{Upper heart rate set at \strong{default = 100 BPM}}

\item{from}{Starting time of plot inclusive (In format \%Y-\%m-\%d)}

\item{to}{Ending time of plot exclusive (In format \%Y-\%m-\%d)}

\item{agg}{Aggregate data by specified time interval}

\item{inter_gap}{Gap allowed between consecutive observations}
}
\value{
Heart rate time series plot for a single subject
}
\description{
The function 'plot_hr' supports time series plotting for both single and multiple subject data
}
\examples{
data(example_heart_1)
plot_hr(example_heart_1)

}
