% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_hr.R
\name{summary_hr}
\alias{summary_hr}
\title{Calculate summary heart rate}
\usage{
summary_hr(data, tz)
}
\arguments{
\item{data}{A DataFrame object with column names "id", "time", "hr".}

\item{tz}{A character string specifying the time zone to be used. System-specific (see \code{\link{as.POSIXct}}), but " " is the current time zone, and "GMT" is UTC (Universal Time, Coordinated). Invalid values are most commonly treated as UTC, on some platforms with a warning
Missing HR values (NA) are automatically excluded from calculations.}
}
\value{
If a dataframe object is passed, then a tibble object with a column for subject id and a column for each of summary values is returned.
'NA' heartrate values are omitted from the calculation of the summary values.
}
\description{
The function summary_hr computes summary statistics for heart rate (HR) data.
It is a wrapper for summary() functions and outputs a tibble object with the subject ID and the following summary values:
Minimum, 1st Quartile, Median, Mean, 3rd Quartile, Maximum, Standard Deviation (SD).
}
\examples{
data(example_heart_1)
summary_hr(example_heart_1)
}
